feat: refactor encoder module
